requirements.txt
Flask==3.1.1
gunicorn
psycopg2-binary
twilio
python-dotenv
